fix: wire Discord approval buttons to real plugin outbound adapter
The sendPayload method was added to the dead legacy adapter in
src/channels/plugins/outbound/discord.ts instead of the actual Discord
plugin loaded by the runtime (extensions/discord/src/channel.ts).
This meant handler.sendPayload was always undefined and the system
fell through to sendText, delivering plain text instead of embeds
with interactive buttons.

Changes:
- Add sendPayload to the real Discord plugin outbound adapter
  (extensions/discord/src/channel.ts) that handles channelData with
  pre-built embeds and components
- Add components support to sendMessageDiscord (send.outbound.ts)
  with a rich-message path for embed/component-only messages
- Update exec-approval-forwarder to build Discord embeds + button
  components at the forwarder level via formatExecApprovalEmbed and
  buildExecApprovalComponents, passing them as channelData.discord
- Revert dead-code sendPayload from src/channels/plugins/outbound/discord.ts
- Update forwarder test assertions to match new channelData shape
