+++ src/discord/send.outbound.ts
@@ -29,6 +29,7 @@ type DiscordSendOpts = {
   replyTo?: string;
   retry?: RetryConfig;
   embeds?: unknown[];
+  components?: unknown[];
 };
 
 export async function sendMessageDiscord(
@@ -53,7 +54,23 @@ export async function sendMessageDiscord(
   const { channelId } = await resolveChannelId(rest, recipient, request);
   let result: { id: string; channel_id: string } | { id: string | null; channel_id: string };
   try {
-    if (opts.mediaUrl) {
+    if (
+      !textWithTables.trim() &&
+      !opts.mediaUrl &&
+      (opts.embeds?.length || opts.components?.length)
+    ) {
+      // Rich message: embeds/components only, no text content required.
+      result = (await request(
+        () =>
+          rest.post(Routes.channelMessages(channelId), {
+            body: {
+              embeds: opts.embeds?.length ? opts.embeds : undefined,
+              components: opts.components?.length ? opts.components : undefined,
+            },
+          }) as Promise<{ id: string; channel_id: string }>,
+        "rich",
+      )) as { id: string; channel_id: string };
+    } else if (opts.mediaUrl) {
       result = await sendDiscordMedia(
         rest,
         channelId,
