+++ src/infra/exec-approval-forwarder.test.ts
@@ -35,11 +35,13 @@ describe("exec approval forwarder", () => {
     await forwarder.handleRequested(baseRequest);
     expect(deliver).toHaveBeenCalledTimes(1);
 
-    // Verify channelData is included in the request delivery
+    // Verify channelData contains pre-built Discord embeds and components
     const requestPayloads = deliver.mock.calls[0][0].payloads;
     expect(requestPayloads[0].channelData).toBeDefined();
-    expect(requestPayloads[0].channelData.discord.execApproval.id).toBe("req-1");
-    expect(requestPayloads[0].channelData.discord.execApproval.command).toBe("echo hello");
+    expect(requestPayloads[0].channelData.discord.embeds).toBeInstanceOf(Array);
+    expect(requestPayloads[0].channelData.discord.embeds.length).toBe(1);
+    expect(requestPayloads[0].channelData.discord.components).toBeInstanceOf(Array);
+    expect(requestPayloads[0].channelData.discord.components.length).toBe(1);
 
     await forwarder.handleResolved({
       id: baseRequest.id,
